refactor: 버블 상태 관리(reducer -> zustand)
