+++ src/components/select/controlSelect/ControlSelect.tsx
@@ -10,13 +10,13 @@ import { useLog } from '@/objects/log/useLog';
 import { useRenderer } from '@/objects/renderer/useRenderer';
 import { useConfigStore } from '@/store/configStore';
 import { useCamera } from '@/objects/camera/useCamera';
-import { useBubble } from '@/objects/bubble/useBubble';
+import { useBubbleStore } from '@/store/bubbleStore';
 
 export const ControlSelect = () => {
     const { cameraView } = useViewStore((state) => state);
     const { mode } = useConfigStore((state) => state);
     const { isUndoAvailable, isRedoAvailable, undo, redo } = useLog();
-    const { setFocusBubblePath } = useBubble();
+    const setFocusBubblePath = useBubbleStore((state) => state.setFocusBubblePath);
     const { reRender } = useRenderer();
     const { updateCameraView } = useCamera();
     // TODO: 투터치 인터페이스 구현(줌/아웃 가능)
