+++ src/util/constant.ts
@@ -1,3 +1,4 @@
 export const MINIMUN_RENDERED_BUBBLE_SIZE = 30;
 export const RENDERED_FONT_SIZE = 22;
 export const BUBBLE_BORDER_WIDTH = 3;
+export const UNNAMED = '제목 없음';
